+++ Changelog.md
@@ -8,6 +8,7 @@
 
 ### Bugfixes
 
+- Make sure the 'To The Dashboard' button is more visible on the onboard/resources page [#275](https://github.com/PrefectHQ/ui/pull/275)
 - Allow label edit on flows created with an older version of core [#273](https://github.com/PrefectHQ/ui/pull/273)
 - Redirect to home when UI is loaded up against a Prefect Server instance with no tenants [#270](https://github.com/PrefectHQ/ui/pull/270)
 - Direct users to the new tenant when they accept an invitation from the accept invitation page [#274](https://github.com/PrefectHQ/ui/pull/274)